Extract GitHub.JWTClient module
